requirements.txt
cs50
Flask
Flask-Session
requests
app.py
templates-folder
styles.css
functions.py
pathology.db
README.md
